requirements.txt
langchain-core 
langchain-azure-ai 
pypdf 
faiss-cpu 
python-dotenv
azure-identity
flask


langchain>=0.2.14
langchain-community>=0.2.14
langchain-text-splitters>=0.2.2
langchain-openai>=0.2.0
faiss-cpu
python-dotenv
flask